Organize folders and add Manager controller. Change 'authentication' from controller to service. Add blanck Nanager controller.
