--- src/main/java/org/fogbowcloud/manager/core/AuthenticationController.java
@@ -1,9 +1,0 @@
-package org.fogbowcloud.manager.core;
-
-public class AuthenticationController {
-	
-	public boolean isAuthorized() {
-		return true;
-	}
-	
-}
