+++ src/main/java/org/fogbowcloud/manager/core/controllers/ApplicationController.java
@@ -1,16 +1,21 @@
-package org.fogbowcloud.manager.core;
+package org.fogbowcloud.manager.core.controllers;
 
 import java.util.Collection;
 
 import org.fogbowcloud.manager.core.models.orders.ComputeOrder;
 import org.fogbowcloud.manager.core.models.orders.NetworkOrder;
 import org.fogbowcloud.manager.core.models.orders.StorageOrder;
+import org.fogbowcloud.manager.core.services.AuthenticationService;
 
 public class ApplicationController {
-
-	//private AuthenticationController authenticationController;
-	//private ManagerController managerController;
+	private AuthenticationService authenticationController;
+	private ManagerController managerController;
 	
+	public ApplicationController(AuthenticationService authenticationController, ManagerController managerController) {
+		this.authenticationController = authenticationController;
+		this.managerController = managerController;
+	}
+
 	public ApplicationController() {}
 
 	public void allocateComputeOrder(ComputeOrder computeOrder) {
@@ -60,5 +65,4 @@ public class ApplicationController {
 	public void removeStorageOrder(Integer id) {
 		
 	}	
-	
 }
