+++ src/main/java/org/fogbowcloud/manager/core/controllers/ManagerController.java
@@ -1,0 +1,5 @@
+package org.fogbowcloud.manager.core.controllers;
+
+public class ManagerController {
+
+}
