+++ src/main/java/org/fogbowcloud/manager/core/services/AuthenticationService.java
@@ -1,0 +1,9 @@
+package org.fogbowcloud.manager.core.services;
+
+public class AuthenticationService {
+	
+	// By pass
+	public boolean isAuthorized() {
+		return true;
+	}
+}
